refactor: simplify HomePage component by integrating gym management logic and enhancing UI structure
